consolidated types into imports handling code
